Create New Bash Script
`set-desktop.sh` replaces both `set-desktop-catalina.sh` and
`set-desktop-big-sur.sh`.

It uses data in the newly created configuration file `options.json` to
validate options passed to it and update the `data` and `preferences`
tables in the `desktoppicture.db` database accordingly.

Can be used for both macOS Catalina and macOS Big Sur.