Change Test for Installed macOS Version
The version numbering convention has changed with the release of Big
Sur where only the major version number is significant for determining
if Big Sur is installed i.e. 11.0, 11.1, 11.2 etc are all versions of
Big Sur.

This is in contrast to earlier macOS versions where both the major and
minor version numbers are significant i.e 10.15 Catalina, 10.14 Mojave
etc.