Structure Changes and Additional Option Configuration Data
- Structure Changes

Within the 'options' array, each 'option' object now includes a 'rows'
array. This array contains a minimum of 1 and a maximum of 2 objects.
Each object defines the value to be inserted into 'value' column of the
'data' table and the value to be inserted into the 'key' column of the
'preferences' table.

The number of objects in the 'rows' array determines the number of rows
to be inserted into the 'data' and 'preferences' table of the
'desktoppicture' database. A single object results in 1 row being inserted
into the 'data' table and 4 rows being inserted into the 'preferences'
table. A total of 2 objects results in 2 rows being inserted into the
'data' table and 8 rows being inserted into the 'preferences' table.

- Additional Option Configuration Data

Option configuration data is now included for images categorised as
Desktop Pictures in System Preferences for macOS Monterey and macOS Big
Sur. The script can set the Desktop picture to one of these images using
either the full path to the image file or its more convenient option
name. This is in response to the location of Apple-supplied Desktop images
changing in macOS Monterey.