Add Test for Python Version
The property list file is parsed by Python's plistlib module.

The version for Python 2.7 is unable to parse binary property list
files and uses the readPlist() function to parse XML-based property list
files.

The version for Python 3 can parse both XML-based and binary property
list files. In addition, the readPlist() function has been deprecated
in favour of the load() function.

All versions of macOS since Mojave 10.14 upto and including Monterey 12
include Python 2.7 as standard and do not ship with Python 3. However,
it's possible that Python 3 has been installed post macOS installation.

Which function and which version of the property list file are used is
now determined by the Python version.