Add File Containing Information on Downloadable Desktop Images
The Apple-supplied file 'com_apple_MobileAsset_DesktopPicture.xml' contains
information required to download Desktop images. It is located in the
'/System/Library/AssetsV2/com_apple_MobileAsset_DesktopPicture', but
doesn't exist until the first Desktop image is downloaded through System
Preferences.

In order to download images, the script first checks for the existance
of this file and if it doesn't exist will use this supplied version.