Add Test for Installed Python Binary
The script checks for and uses a Python 3 binary if one exists, if not it
uses an available Python 2 binary. The exception is when the installed
macOS version is Catalina (10.15). In this instance, Python 2 is used
as `/usr/bin/python3` is merely a stub for installing the command line
developer tools that include Python 3.